Split storage spoke dialogs to a separate module
Splits also the glade file accordingly.
